+++ spec/fixtures/monorepo/admin_app/app.rb
@@ -1,0 +1,9 @@
+class AdminApp
+  def load
+    return :app
+  end
+
+  def render
+    return :render
+  end
+end