+++ spec/fixtures/monorepo/app/db/migrate/202511251234567_create_foos.rb
@@ -1,0 +1,5 @@
+class CreateFoos < Migration
+  def change
+    create_foos!
+  end
+end